# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Suivi Académique",
        "description": "Documentation de l'API pour le suivi académique",
        "contact": {
            "email": "contact@suivi-academique.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur API"
        }
    ],
    "paths": {
...
        "/api/niveaux": {
            "get": {
                "tags": [
                    "Niveaux"
                ],
                "summary": "Obtenir la liste des niveaux",
                "description": "Retourne la liste de tous les niveaux avec leurs UE associées",
                "operationId": "getNiveauxList",
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "code_niveau": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "label_niveau": {
                                                "type": "string",
                                                "example": "Niveau 1"
                                            },
                                            "description_niveau": {
                                                "type": "string",
                                                "example": "Premier niveau d'étude"
                                            },
                                            "code_filiere": {
                                                "type": "string",
                                                "example": "INFO-001"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "ue": {
                                                "description": "Liste des UE associées à ce niveau",
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "code_ue": {
                                                            "type": "string",
                                                            "example": "UE-101"
                                                        },
                                                        "label_ue": {
                                                            "type": "string",
                                                            "example": "Programmation"
                                                        },
                                                        "description_ue": {
                                                            "type": "string",
                                                            "example": "Introduction à la programmation"
                                                        },
                                                        "code_niveau": {
                                                            "type": "integer",
                                                            "example": "1"
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        },
                                                        "updated_at": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Niveaux"
                ],
                "summary": "Créer un niveau",
                "operationId": "storeNiveau",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code_filiere",
                                    "label_niveau",
                                    "description_niveau"
                                ],
                                "properties": {
                                    "code_filiere": {
                                        "type": "string"
                                    },
                                    "label_niveau": {
                                        "type": "string"
                                    },
                                    "description_niveau": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Niveau créé"
                    },
                    "422": {
                        "description": "Erreur validation"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/api/niveaux/{codeNiveau}": {
            "get": {
                "tags": [
                    "Niveaux"
                ],
                "summary": "Obtenir une niveau par son code",
                "description": "Retourne les informations d'une niveau spécifique",
                "operationId": "showNiveau",
                "parameters": [
                    {
                        "name": "codeNiveau",
                        "in": "path",
                        "description": "Code unique de la niveau",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Niveau trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code_niveau": {
                                            "type": "integer"
                                        },
                                        "label_niveau": {
                                            "type": "string"
                                        },
                                        "description_niveau": {
                                            "type": "string"
                                        },
                                        "code_filiere": {
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "ue": {
                                            "description": "Liste des UE associées à ce niveau",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "code_ue": {
                                                        "type": "string",
                                                        "example": "UE-101"
                                                    },
                                                    "label_ue": {
                                                        "type": "string",
                                                        "example": "Programmation"
                                                    },
                                                    "description_ue": {
                                                        "type": "string",
                                                        "example": "Introduction à la programmation"
                                                    },
                                                    "code_niveau": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Niveau non trouvée"
                    }
                }
            },
            "put": {
                "tags": [
                    "Niveaux"
                ],
                "summary": "Mettre à jour un niveau",
                "description": "Met à jour les informations d'une niveau existant",
                "operationId": "updateNiveau",
                "parameters": [
                    {
                        "name": "codeNiveau",
                        "in": "path",
                        "description": "Code unique du niveau",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code_filiere",
                                    "label_filiere",
                                    "description_filiere"
                                ],
                                "properties": {
                                    "code_filiere": {
                                        "type": "string"
                                    },
                                    "label_niveau": {
                                        "type": "string"
                                    },
                                    "description_niveau": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Niveau mise à jour"
                    },
                    "404": {
                        "description": "Niveau non trouvé"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Niveaux"
                ],
                "summary": "Supprimer un niveau",
                "operationId": "deleteNiveau",
                "parameters": [
                    {
                        "name": "codeNiveau",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Niveau supprimée"
                    },
                    "404": {
                        "description": "Niveau non trouvé"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
...
    },
...
}